% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groan.R
\name{groan_id}
\alias{groan_id}
\title{Function to return a specific dad joke}
\usage{
groan_id(joke_id)
}
\arguments{
\item{joke_id}{A specific dad joke ID to return}
}
\value{
Returns a two item list with the first item holding the joke as a 
        character and the second item containing the joke id as a character.
}
\description{
This function returns a specific dad joke from 
\url{https://icanhazdadjoke.com} using its dad joke ID.
}
\examples{
groan_id("GlGBIY0wAAd")
}
