% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groan.R
\name{groan_image}
\alias{groan_image}
\title{Function to return a specific dad joke as an image}
\usage{
groan_image(joke_id)
}
\arguments{
\item{joke_id}{A specific dad joke ID to return}

\item{image}{A logical to indicate if an image of a specific joke ID should 
be returned instead of text.}
}
\value{
Returns a png array from png::readPNG.
}
\description{
This function returns a specific dad joke as a png from 
\url{https://icanhazdadjoke.com} using its dad joke ID.
}
\examples{
joke_png <- groan_image("GlGBIY0wAAd")
}
