% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groan.R
\name{groan_search}
\alias{groan_search}
\title{Function to search and return dad jokes with specific terms}
\usage{
groan_search(term, sting = TRUE)
}
\arguments{
\item{term}{A term to search \url{https://icanhazdadjoke.com}}

\item{sting}{Plays a joke sting after the joke.  Sound from
\url{https://archive.org/details/Rimshot_254}}
}
\value{
Returns a data frame of joke and joke id for jokes that contain the 
        search term.
}
\description{
This function returns a specific dad joke as a png from 
\url{https://icanhazdadjoke.com} using its dad joke ID.
}
\examples{
groan()
}
