% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomjoke.R
\name{randomjoke}
\alias{randomjoke}
\title{Function to return random joke from package functions}
\usage{
randomjoke()
}
\description{
This function chooses a joke to return, calling either dadjoke() or chuckjoke() at random.
}
\examples{
randomjoke()
}
