% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chuckjoke.R
\name{chuckjoke}
\alias{chuckjoke}
\title{Function to return random Chuck Norris joke}
\usage{
chuckjoke(explicit = FALSE)
}
\description{
This function returns a random Chuck Norris joke from \url{https://icndb.com}
}
\examples{
chuckjoke()
}
