% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groan.R
\name{groan}
\alias{groan}
\title{Function to return random dad joke}
\usage{
groan(rimshot = TRUE)
}
\arguments{
\item{rimshot}{Plays a rimshot after the joke.  Sound from
\url{https://archive.org/details/Rimshot_254}}
}
\description{
This function returns a random dad joke from \url{https://icanhazdadjoke.com}
}
\examples{
groan()
}
