% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groan.R
\name{groan}
\alias{groan}
\title{Function to return random dad joke}
\usage{
groan(sting = TRUE)
}
\arguments{
\item{sting}{Plays a joke sting after the joke.  Sound from
\url{https://archive.org/details/Rimshot_254}}
}
\value{
Invisibly returns a two item list with the first item holding the
        joke as a character and the second item containing the joke id as a
        character.
}
\description{
This function returns a random dad joke(s) from 
\url{https://icanhazdadjoke.com}. as a message and invisibly as a character.
}
\examples{
groan(sting = FALSE)
}
