% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groan.R
\name{groan_id}
\alias{groan_id}
\title{Function to return a specific dad joke}
\usage{
groan_id(joke_id, sting = TRUE)
}
\arguments{
\item{joke_id}{A specific dad joke ID to return}

\item{sting}{Plays a joke sting after the joke.  Sound from
\url{https://archive.org/details/Rimshot_254}}

\item{image}{A logical to indicate if an image of a specific joke ID should 
be returned instead of text.}
}
\value{
Returns a two item list with the first item holding the joke as a 
        character and the second item containing the joke id as a character.
}
\description{
This function returns a specific dad joke from 
\url{https://icanhazdadjoke.com} using its dad joke ID.
}
\examples{
groan()
}
